Refactor of the three Manager classes:
* Split cleanup() into metadata cleanup (push to MetadataManager) and physical object clean up
* Remove object_engine from MetadataManager (only concerns itself with metadata_engine)
* Push object filtering/deletions into FragmentManager, make filter_objects public
